Cone position in CT images
Changing the method we I use to find user clicked indices for cone construction. I use binary Image with clicked coordinates. Rotating this binary image by the same angle but the negative direction will result in finding those indices in the not rotated image. In order to find the exact place of the indices after the interpolation of the rotation function I use cv2 connected components function which finds clusters of binary images. in each cluster I then find the maximum value and take is indices to be set as the index for the cone.